Move GradingScheme types to separate files (#3270)
affects: Parent, Student

[ignore-commit-lint]
