+++ Core/Core/Features/Grades/Model/Entities/GradingScheme.swift
@@ -32,103 +32,3 @@ public extension GradingScheme {
         return entries.first { normalizedScore >= $0.value }?.name
     }
 }
-
-public struct PercentageBasedGradingScheme: GradingScheme {
-
-    public let entries: [GradingSchemeEntry]
-
-    fileprivate init(entries: [GradingSchemeEntry]) {
-        self.entries = entries
-    }
-
-    public func formattedScore(from value: Double) -> String? {
-        Self.percentFormatter.string(from: NSNumber(value: value))
-    }
-
-    private static let percentFormatter: NumberFormatter = {
-        let formatter = NumberFormatter()
-        formatter.numberStyle = .percent
-        formatter.decimalSeparator = "."
-        formatter.multiplier = 1
-        formatter.maximumFractionDigits = 3
-        formatter.roundingMode = .down
-        return formatter
-    }()
-}
-
-public struct PointsBasedGradingScheme: GradingScheme {
-
-    public let scaleFactor: Double
-    public let entries: [GradingSchemeEntry]
-
-    fileprivate init(scaleFactor: Double, entries: [GradingSchemeEntry]) {
-        self.scaleFactor = scaleFactor
-        self.entries = entries
-    }
-
-    public func formattedScore(from value: Double) -> String? {
-        let normalizedScore = value / 100.0
-        let number = NSNumber(value: normalizedScore * scaleFactor)
-        return Self.pointsFormatter.string(from: number)
-    }
-
-    private static let pointsFormatter: NumberFormatter = {
-        let formatter = NumberFormatter()
-        formatter.numberStyle = .decimal
-        formatter.multiplier = 1
-        formatter.maximumFractionDigits = 2
-        formatter.roundingMode = .halfEven
-        return formatter
-    }()
-}
-
-// MARK: - Schemes for Previews & Testing
-
-#if DEBUG
-
-public extension PercentageBasedGradingScheme {
-    static var `default`: Self {
-        PercentageBasedGradingScheme(entries: [])
-    }
-
-    static func make(entries: [GradingSchemeEntry]) -> Self {
-        PercentageBasedGradingScheme(entries: entries)
-    }
-}
-
-public extension PointsBasedGradingScheme {
-    static var `default`: Self {
-        PointsBasedGradingScheme(scaleFactor: 4, entries: [])
-    }
-
-    static func make(scaleFactor: Double, entries: [GradingSchemeEntry]) -> Self {
-        PointsBasedGradingScheme(scaleFactor: scaleFactor, entries: entries)
-    }
-}
-
-public extension GradingScheme where Self == PercentageBasedGradingScheme {
-    static var defaultPercentageBased: GradingScheme {
-        PercentageBasedGradingScheme.default
-    }
-}
-
-public extension GradingScheme where Self == PointsBasedGradingScheme {
-    static var defaultPointsBased: GradingScheme {
-        PointsBasedGradingScheme.default
-    }
-}
-
-#endif
-
-// MARK: - As Course's property
-
-extension Course {
-
-    public var gradingScheme: GradingScheme {
-        if pointsBasedGradingScheme {
-            return PointsBasedGradingScheme(scaleFactor: scalingFactor, entries: gradingSchemeEntries)
-        } else {
-            return PercentageBasedGradingScheme(entries: gradingSchemeEntries)
-        }
-    }
-}
